Implement goal tracking (#2)
* Update daisyui

* Add goals model, actions and service

* Create goals page

* Add EditGoalButton and CalendarChart to goal page

* Improve CalendarChart and add goal item edit

* Minor goal styles adjustment

* Display length above textarea

* Create rechart goal bar chart

* Modify goal page layout

* Block non-admin users from making public blogs

* Formalize goal completion flow

* Update goal creation, metrics calculation, and UI components

* Add goal status tracking and completion logic

* Notify client about goal completion

* Implement goal deletion with confirmation

* Add confirmation dialog for blog deletion

* Update docs

* Add goal table name mapping

* Versioning untracked file

* Remove obsolete classname

* Add messages for no page results

+++ src/app/blogs/EditBlogForm.tsx
@@ -33,9 +33,14 @@ type FormValues = {
 type EditBlogFormProps = {
   blog?: BlogDTO;
   tagCounts: TagWithCount[];
+  canCreatePublicBlog: boolean;
 };
 
-export default function EditBlogForm({ blog, tagCounts }: EditBlogFormProps) {
+export default function EditBlogForm({
+  blog,
+  tagCounts,
+  canCreatePublicBlog,
+}: EditBlogFormProps) {
   const router = useRouter();
   const blobManagerRef = React.useRef(BlobManager.getInstance());
 
@@ -60,7 +65,7 @@ export default function EditBlogForm({ blog, tagCounts }: EditBlogFormProps) {
   const onSubmit: SubmitHandler<FormValues> = async (data) => {
     const { content, images: contentImages } = await uploadNewContentImages(
       data.content,
-      blobManagerRef.current
+      blobManagerRef.current,
     );
 
     const [coverImage] = coverImagePreview
@@ -69,7 +74,7 @@ export default function EditBlogForm({ blog, tagCounts }: EditBlogFormProps) {
 
     const prevImages = blog?.images || [];
     const prevImagesStillInContent = prevImages.filter((img) =>
-      content.match(new RegExp(img.url, "g"))
+      content.match(new RegExp(img.url, "g")),
     );
 
     const body = {
@@ -102,6 +107,13 @@ export default function EditBlogForm({ blog, tagCounts }: EditBlogFormProps) {
 
   const onDelete = () => {
     if (!blog?.id) return;
+    if (
+      !window.confirm(
+        "Are you sure you want to delete this blog? This action cannot be undone.",
+      )
+    ) {
+      return;
+    }
     setIsDeleting(true);
     deleteBlog(blog.id)
       .then((res) => {
@@ -187,8 +199,10 @@ export default function EditBlogForm({ blog, tagCounts }: EditBlogFormProps) {
               className="w-full"
               required
             >
-              <option value={BlogVisibility.PUBLIC}>🌍 Everyone</option>
-              <option value={BlogVisibility.FRIENDS}>🧑‍🤝‍🧑 Friends only</option>
+              {canCreatePublicBlog && (
+                <option value={BlogVisibility.PUBLIC}>🌍 Everyone</option>
+              )}
+              <option value={BlogVisibility.FRIENDS}>🧑‍🤝‍🧑 Friends</option>
               <option value={BlogVisibility.PRIVATE}>🔒 Just me</option>
             </Select>
             <Controller
